backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
fastapi-cors
alpha-vantage
scikit-learn
joblib
python-dateutil
pandas
openpyxl
fpdf2
matplotlib
passlib[bcrypt]
python-jose[cryptography]